refactor: extract session helpers, remove stripAnsi/truncate duplicates
